requirements.txt
# ===========================
# Backend Dependencies
# ===========================

# FastAPI & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
starlette>=0.35.0

# Database
sqlalchemy>=2.0.25
pymysql>=1.1.0
cryptography>=42.0.0

# Authentication
PyJWT[crypto]>=2.8.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# ===========================
# RAG Service Dependencies
# ===========================

# LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
langgraph>=0.0.20

# LLM & Embeddings
openai>=1.6.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector DB
chromadb==1.5.0

# RAG Components
rank-bm25>=0.2.2

# Web & HTTP
httpx>=0.26.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Search
tavily-python>=0.3.0

# Document Generation
reportlab>=4.0.0
python-docx>=1.0.0

# ===========================
# AWS
# ===========================
boto3>=1.26.0

# ===========================
# Shared Utilities
# ===========================
python-dotenv>=1.0.0
tqdm>=4.66.0

# ===========================
# Testing & Evaluation
# ===========================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
ragas>=0.1.0
datasets>=2.14.0
